requirements_base.txt
Django
django-extensions
ipython
dj-database-url
pytz
djangorestframework
markdown
django-filter
django-oauth-toolkit